requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
Flask==3.0.0
selenium==4.15.2
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
gunicorn==21.2.0



